Use watchdog to reset and jump to application_jump_check
1. Using the watchdog to reset and jump to bootloader is the
   recommended method my LUFA:

(From It's software bootloader documentation)
> By using the watchdog to physically reset the controller, it is
ensured that all system hardware is completely reset to their
defaults before the bootloader is run. This is important;
since bootloaders are written to occupy a very limited space,
they usually make assumptions about the register states based on
the default values after a hard-reset of the chip.

2. Jump to application jump check rather than the bootloader
directly, LUFA expects this when bootloading via a watchdog reset
eventhough this isn't documented anywhere.